offload drawing functions
